Update README.md
Modify the Apartment class so that it holds an aptNumber (apartment number), bedrooms (number of bedrooms), baths (number of baths), and rent (rent amount). Create a constructor that accepts values for each data field. Also create a get method for each field.

Modify the TestApartments application so that it creates at least five Apartment objects. Then prompt a user to enter a minimum number of bedrooms required, a minimum number of baths required, and a maximum rent the user is willing to pay. Display data for all the Apartment objects that meet the user’s criteria or No apartments met your criteria if no such apartments are available.